........
r60712 | tilghman | 2007-04-08 09:12:00 -0500 (Sun, 08 Apr 2007) | 2 lines

Fix --enable-dev-mode

........


git-svn-id: https://origsvn.digium.com/svn/asterisk/trunk@60714 65c4cc65-6c06-0410-ace0-fbb531ad65f3

+++ main/http.c
@@ -403,7 +403,7 @@ static void post_raw(struct mm_mimepart *part, const char *post_dir, const char 
 	body_len = mm_mimepart_getlength(part);
 
 	if (option_debug)
-		ast_log(LOG_DEBUG, "Body length is %ld\n", body_len);
+		ast_log(LOG_DEBUG, "Body length is %ld\n", (long int)body_len);
 
 	fwrite(body, 1, body_len, f);
